* In order to make it link without TBB it was also rrequired to
   merge miagsl into miacore

+++ mia/core/labelmap.hh
@@ -1,7 +1,7 @@
 /* -*- mia-c++  -*-
  *
  * This file is part of MIA - a toolbox for medical image analysis 
- * Copyright (c) Leipzig, Madrid 1999-2014 Gert Wollny
+ * Copyright (c) Leipzig, Madrid 1999-2015 Gert Wollny
  *
  * MIA is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -33,7 +33,7 @@ NS_MIA_BEGIN
    \ingroup misc
    \brief A simple class to add loadind and storeing to a map of labels. 
 */
-class EXPORT_CORE CLabelMap: public std::map<unsigned short,  unsigned short> {
+class EXPORT_CORE CLabelMap: public std::map<unsigned int,  unsigned int> {
 public: 
 	CLabelMap() = default;
